Reading and training from actual wrangled data
I made a quick merge of the-movies dataset and did some preliminary cleaning of the data - with numpy and pandas we can successfully load in csv's to a numpy structured array. We'll need to do quite a bit of cleaning of the data though.

Also, this is just a tester - it uses the same data for the training/test/evaluating - we should instead randomize the rows of the array and use subsets for each of these datasets.